+++ source/index.ts
@@ -4,5 +4,5 @@
 // Export all the types as named exports.
 export * from './types.js'
 
-// Export the slowDown function as well.
-export { default, default as slowDown } from './express-slow-down.js'
+// Export the `slowDown` function as well.
+export { default, default as slowDown } from './slow-down.js'
